add only-content option to getCanvasRectangleFromElement
