+++ editor/src/components/editor/store/store-deep-equality-instances.ts
@@ -1478,17 +1478,6 @@ export function SidesKeepDeepEquality(
   }
 }
 
-export function TextBoundsKeepEquality(
-  oldSides: DOMRect,
-  newSides: DOMRect,
-): KeepDeepEqualityResult<DOMRect> {
-  if (oldSides.width === newSides.width && oldSides.height === newSides.height) {
-    return keepDeepEqualityResult(oldSides, true)
-  } else {
-    return keepDeepEqualityResult(newSides, false)
-  }
-}
-
 const SameFileOriginKeepDeepEquality: KeepDeepEqualityCall<SameFileOrigin> = combine2EqualityCalls(
   (i) => i.filePath,
   createCallWithTripleEquals(),
@@ -1595,7 +1584,7 @@ export function SpecialSizeMeasurementsKeepDeepEquality(): KeepDeepEqualityCall<
     const fontWeightEquals = oldSize.fontWeight === newSize.fontWeight
     const fontStyleEquals = oldSize.fontStyle === newSize.fontStyle
     const textDecorationLineEquals = oldSize.textDecorationLine === newSize.textDecorationLine
-    const textBoundsEqual = nullableDeepEquality(TextBoundsKeepEquality)(
+    const textBoundsEqual = nullableDeepEquality(CanvasRectangleKeepDeepEquality)(
       oldSize.textBounds,
       newSize.textBounds,
     ).areEqual
