+++ editor/src/core/shared/dom-utils.ts
@@ -1,8 +1,16 @@
 import type { ReactDOM } from 'react'
 import type { CanvasRectangle } from './math-utils'
-import { boundingRectangle, canvasRectangle, roundToNearestHalf, scaleRect } from './math-utils'
+import {
+  boundingRectangle,
+  canvasRectangle,
+  roundToNearestHalf,
+  scaleRect,
+  zeroCanvasRect,
+} from './math-utils'
 import { URL_HASH } from '../../common/env-vars'
 import { blockLevelHtmlElements, inlineHtmlElements } from '../../utils/html-elements'
+import { parseCSSPx } from '../../components/inspector/common/css-utils'
+import { isRight } from './either'
 
 export const intrinsicHTMLElementNames: Array<keyof ReactDOM> = [
   'a',
@@ -227,8 +235,12 @@ export function setDOMAttribute(element: Element, attributeName: string, value: 
 export function getCanvasRectangleFromElement(
   element: HTMLElement,
   canvasScale: number,
-  withContent: 'without-content' | 'with-content',
+  withContent: 'without-content' | 'with-content' | 'only-content',
 ): CanvasRectangle {
+  if (withContent === 'only-content') {
+    return getContentBounds(element, canvasScale) ?? zeroCanvasRect
+  }
+
   const scale = canvasScale < 1 ? 1 / canvasScale : 1
 
   const domRectToScaledCanvasRectangle = (rect: DOMRect) => {
@@ -257,6 +269,41 @@ export function getCanvasRectangleFromElement(
     typeof range.getBoundingClientRect === 'function' ? range.getBoundingClientRect() : boundingRect
 
   return boundingRectangle(elementRect, domRectToScaledCanvasRectangle(rangeBounding))
+}
+
+function getContentBounds(element: HTMLElement, scale: number): CanvasRectangle | null {
+  const range = document.createRange()
+  range.selectNodeContents(element)
+  if (range.getBoundingClientRect == null) {
+    // This can be undefined in tests
+    return null
+  }
+  const contentBounds = range.getBoundingClientRect()
+
+  const computedStyle = window.getComputedStyle(element)
+  function maybeValueFromComputedStyle(property: string): number {
+    const parsed = parseCSSPx(property)
+    return isRight(parsed) ? parsed.value.value : 0
+  }
+
+  const ratio = scale < 1 ? scale : 1
+  return canvasRectangle({
+    ...contentBounds,
+    width:
+      (contentBounds.width +
+        maybeValueFromComputedStyle(computedStyle.paddingLeft) +
+        maybeValueFromComputedStyle(computedStyle.paddingRight) +
+        maybeValueFromComputedStyle(computedStyle.marginLeft) +
+        maybeValueFromComputedStyle(computedStyle.marginRight)) /
+      ratio,
+    height:
+      (contentBounds.height +
+        maybeValueFromComputedStyle(computedStyle.paddingTop) +
+        maybeValueFromComputedStyle(computedStyle.paddingBottom) +
+        maybeValueFromComputedStyle(computedStyle.marginTop) +
+        maybeValueFromComputedStyle(computedStyle.marginBottom)) /
+      ratio,
+  })
 }
 
 export function addStyleSheetToPage(url: string, shouldAppendHash: boolean = true): void {
