+++ editor/src/core/shared/utils.ts
@@ -13,7 +13,7 @@ export const CanvasContextMenuPortalTargetID = 'canvas-contextmenu-portal-target
 export const RETURN_TO_PREPEND = 'return '
 
 // eslint-disable-next-line @typescript-eslint/no-empty-function
-export const NO_OP = () => {}
+export function NO_OP() {}
 
 export type ValueOf<T> = T[keyof T]
